feat: embrace workspaces
