+++ Cargo.toml
@@ -9,3 +9,43 @@ members = [
 # Only check / build main crates by default (check all with `--workspace`)
 default-members = ["insim", "insim_*"]
 resolver = "2"
+
+[workspace.package]
+version = "1.1.1"
+authors = ["Karl Southern <karl@theangryangel.co.uk>"]
+edition = "2021"
+rust-version = "1.66"
+repository = "https://github.com/theangryangel/insim.rs"
+license = "MIT"
+homepage = "https://github.com/theangryangel/insim.rs"
+
+[workspace.lints.rust]
+unsafe_code = "deny"
+missing_docs = "deny"
+unused_crate_dependencies = "deny"
+unused_results = "deny"
+missing_debug_implementations = "deny"
+
+[workspace.lints.rustdoc]
+broken_intra_doc_links = "deny"
+
+[workspace.dependencies]
+async-trait = { version = "0.1.68" }
+binrw = "0.13.3"
+bitflags = "2.4.0"
+bytes = "1.4.0"
+clap = { version = "4.2" }
+encoding_rs = { version = "0.8.32" }
+from_variants = "1.0.2"
+futures-util = { version = "0.3.28" }
+if_chain = "1.0.2"
+indexmap = "2.1.0"
+itertools = { version = "0.12.0" }
+once_cell = "1.17.1"
+serde = { version = "1.0.188" }
+thiserror = "1.0.40"
+tokio = { version = "1.11.0" }
+tokio-test = "0.4.0"
+tokio-tungstenite = { version = "0.21.0" }
+tracing = "0.1.37"
+tracing-subscriber = { version = "0.3.17" }
