+++ examples/low-level-async/Cargo.toml
@@ -1,13 +1,15 @@
 [package]
 name = "low-level-async"
-version = "0.1.0"
-edition = "2021"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
 publish = false
 
 [dependencies]
-clap = { version = "4.2", features = ["derive"] }
-if_chain = "1"
+clap = { workspace = true, features = ["derive"] }
+if_chain = { workspace = true }
 insim = { path = "../../insim" }
-tokio = { version = "1.11.0", features = ["full"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tokio = { workspace = true, features = ["full"] }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true, features = ["env-filter"] }
