+++ examples/simple-blocking/Cargo.toml
@@ -1,12 +1,14 @@
 [package]
 name = "simple-blocking"
-version = "0.1.0"
-edition = "2021"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
 publish = false
 
 [dependencies]
-clap = { version = "4.2", features = ["derive"] }
-if_chain = "1"
+clap = { workspace = true, features = ["derive"] }
+if_chain = { workspace = true }
 insim = { path = "../../insim", default-features = false, features = ["blocking"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true, features = ["env-filter"] }
