+++ insim/src/lib.rs
@@ -1,10 +1,4 @@
 #![doc = include_str!("../README.md")]
-#![deny(
-    missing_docs,
-    rustdoc::broken_intra_doc_links,
-    missing_debug_implementations,
-    unused_results
-)]
 #![cfg_attr(test, deny(warnings, unreachable_pub))]
 
 use std::net::SocketAddr;
