+++ insim_core/src/lib.rs
@@ -1,4 +1,3 @@
-#![deny(unused_crate_dependencies)]
 #![doc = include_str!("../README.md")]
 
 pub mod duration;
