+++ insim_core/src/string/colours.rs
@@ -1,5 +1,8 @@
+//! Utilities for working with colours from Insim.
+
 use super::MARKER;
 
+/// Supported colour codes within LFS
 pub const COLOUR_SEQUENCES: &[char] = &['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
 
 /// Strip LFS colours
@@ -24,4 +27,19 @@ pub fn strip(input: &str) -> String {
     }
 
     output
+}
+
+#[test]
+fn test_strip_colours_only() {
+    assert_eq!(strip("^1^2^3^4^5^6^7^8^9"), "");
+}
+
+#[test]
+fn test_strip_colours() {
+    assert_eq!(strip("^1234^56789"), "2346789");
+}
+
+#[test]
+fn test_strip_colours_escaped() {
+    assert_eq!(strip("^^1234^56789"), "^^12346789");
 }
