+++ insim_core/src/string/mod.rs
@@ -8,8 +8,10 @@ use if_chain::if_chain;
 pub mod codepages;
 pub mod colours;
 
+/// Escape and special character prefix
 pub const MARKER: u8 = b'^';
 
+/// Special character escape sequences
 pub const ESCAPE_SEQUENCES: &[(u8, u8)] = &[
     (b'v', b'|'),
     (b'a', b'*'),
@@ -131,6 +133,8 @@ pub fn escape(input: Cow<str>) -> Cow<str> {
 
 use binrw::{helpers::until_eof, BinRead, BinWrite};
 
+#[doc(hidden)]
+#[allow(missing_docs)]
 #[binrw::writer(writer, endian)]
 pub fn binrw_write_codepage_string<const SIZE: usize>(
     input: &String,
@@ -166,6 +170,8 @@ pub fn binrw_write_codepage_string<const SIZE: usize>(
     Ok(())
 }
 
+#[doc(hidden)]
+#[allow(missing_docs)]
 #[binrw::parser(reader, endian)]
 pub fn binrw_parse_codepage_string<const SIZE: usize>(raw: bool) -> binrw::BinResult<String> {
     <[u8; SIZE]>::read_options(reader, endian, ()).map(|bytes| {
@@ -177,6 +183,8 @@ pub fn binrw_parse_codepage_string<const SIZE: usize>(raw: bool) -> binrw::BinRe
     })?
 }
 
+#[doc(hidden)]
+#[allow(missing_docs)]
 #[binrw::parser(reader, endian)]
 pub fn binrw_parse_codepage_string_until_eof(raw: bool) -> binrw::BinResult<String> {
     until_eof(reader, endian, ()).map(|bytes: Vec<u8>| {
@@ -186,4 +194,46 @@ pub fn binrw_parse_codepage_string_until_eof(raw: bool) -> binrw::BinResult<Stri
             Ok(codepages::to_lossy_string(&bytes).to_string())
         }
     })?
+}
+
+#[test]
+fn test_escaping_and_unescaping() {
+    let original: Cow<str> = "^|*:\\/?\"<>#123^945".into();
+
+    let escaped = escape(original.clone());
+    let unescaped = unescape(escaped.clone());
+
+    assert_eq!(escaped, "^^^v^a^c^d^s^q^t^l^r^h123^945");
+    assert_eq!(unescaped, original);
+}
+
+#[test]
+fn test_codepage_hello_world() {
+    let output = codepages::to_lossy_bytes("Hello");
+
+    assert_eq!(output, "Hello".as_bytes(),);
+}
+
+// sample utf-8 strings from https://www.cl.cam.ac.uk/~mgk25/ucs/examples/quickbrown.txt
+
+#[test]
+fn test_codepage_to_hungarian() {
+    // flood-proof mirror-drilling machine
+    let as_bytes = codepages::to_lossy_bytes("Árvíztűrő tükörfúrógép");
+
+    assert_eq!(
+        codepages::to_lossy_string(&as_bytes),
+        "Árvízt?r? tükörfúrógép",
+    );
+}
+
+#[test]
+fn test_codepage_to_mixed() {
+    // flood-proof mirror-drilling machine
+    let as_bytes = codepages::to_lossy_bytes("TEST Árvíztűrő tükörfúrógép");
+
+    assert_eq!(
+        codepages::to_lossy_string(&as_bytes),
+        "TEST Árvízt?r? tükörfúrógép",
+    );
 }
