+++ insim_pth/src/lib.rs
@@ -1,5 +1,3 @@
-#![deny(unused_crate_dependencies)]
-
 //! # insim_pth
 //!
 //! Parse a Live for Speed pth (path) file.
@@ -19,6 +17,8 @@
 //! there is approximately 0.2 seconds of time between passing one node and the next,
 //! when you are "driving at a reasonable speed".
 
+#[cfg(test)]
+use std::io::{Cursor, Read, Seek, SeekFrom};
 use std::{
     fs::{self, File},
     io::ErrorKind,
@@ -34,6 +34,7 @@ use thiserror::Error;
 
 #[non_exhaustive]
 #[derive(Error, Debug)]
+#[allow(missing_docs)]
 pub enum Error {
     #[error("IO Error: {kind}: {message}")]
     IO { kind: ErrorKind, message: String },
@@ -55,7 +56,10 @@ impl From<std::io::Error> for Error {
 #[derive(Debug, Copy, Clone, Default, PartialEq)]
 #[binrw]
 pub struct Limit {
+    /// Left track limit
     pub left: f32,
+
+    /// Right track limit
     pub right: f32,
 }
 
@@ -135,15 +139,19 @@ impl Node {
 #[derive(Debug, Default, PartialEq)]
 /// PTH file
 pub struct Pth {
+    /// File format version
     pub version: u8,
+    /// File format revision
     pub revision: u8,
 
     #[bw(calc = nodes.len() as i32)]
     num_nodes: i32,
 
+    /// Which node is the finishing line
     pub finish_line_node: i32,
 
     #[br(count = num_nodes)]
+    /// A list of nodes
     pub nodes: Vec<Node>,
 }
 
@@ -166,4 +174,51 @@ impl Pth {
 
         Self::from_file(&mut input)
     }
+}
+
+#[cfg(test)]
+fn assert_valid_as1_pth(p: &Pth) {
+    assert_eq!(p.version, 0);
+    assert_eq!(p.revision, 0);
+    assert_eq!(p.finish_line_node, 250);
+}
+
+#[test]
+fn test_pth_decode_from_pathbuf() {
+    let path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("./tests/AS1.pth");
+    let p = Pth::from_pathbuf(&path).expect("Expected PTH file to be parsed");
+
+    assert_valid_as1_pth(&p)
+}
+
+#[test]
+fn test_pth_decode_from_file() {
+    let path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("./tests/AS1.pth");
+    let mut file = File::open(path).expect("Expected Autocross_3DH.smx to exist");
+    let p = Pth::from_file(&mut file).expect("Expected PTH file to be parsed");
+
+    let pos = file.stream_position().unwrap();
+    let end = file.seek(SeekFrom::End(0)).unwrap();
+
+    assert_eq!(pos, end, "Expected the whole file to be completely read");
+
+    assert_valid_as1_pth(&p)
+}
+
+#[test]
+fn test_pth_encode() {
+    let path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("./tests/AS1.pth");
+    let p = Pth::from_pathbuf(&path).expect("Expected SMX file to be parsed");
+
+    let mut file = File::open(path).expect("Expected AS1.pth to exist");
+    let mut raw: Vec<u8> = Vec::new();
+    let _ = file
+        .read_to_end(&mut raw)
+        .expect("Expected to read whole file");
+
+    let mut writer = Cursor::new(Vec::new());
+    binrw::BinWrite::write(&p, &mut writer).expect("Expected to write the whole file");
+
+    let inner = writer.into_inner();
+    assert_eq!(inner, raw);
 }
